Make MarkdownTable an inner class of Markdown
